requirements.txt
# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data Processing
pandas==2.1.4
numpy==1.26.2

# ETL Tools
click==8.1.7
python-dotenv==1.0.0
tqdm==4.66.1

# Utilities
requests==2.31.0
pyyaml==6.0.1
loguru==0.7.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0